handle optional subscriber if exist
