+++ tests/Type/Symfony/ExtensionTest.php
@@ -14,6 +14,13 @@ class ExtensionTest extends TypeInferenceTestCase
 	/** @return mixed[] */
 	public function dataFileAsserts(): iterable
 	{
+		if (class_exists('Symfony\Component\Messenger\Handler\MessageSubscriberInterface')) {
+			// todo temporary check
+			die('test if case is triggered');
+
+			yield from $this->gatherAssertTypes(__DIR__ . '/data/messenger_handle_trait_with_subscriber.php');
+		}
+
 		yield from $this->gatherAssertTypes(__DIR__ . '/data/messenger_handle_trait.php');
 		yield from $this->gatherAssertTypes(__DIR__ . '/data/envelope_all.php');
 		yield from $this->gatherAssertTypes(__DIR__ . '/data/header_bag_get.php');
